requirements.txt
chromadb
crewai 
sentence-transformers 
psycopg2-binary